+++ tests/iamnotroot.c
@@ -18,17 +18,6 @@ int __getdebug_fd()
 	return STDERR_FILENO;
 }
 
-int __path_setenv(const char *root, const char *name, const char *value,
-		 int overwrite)
-{
-	(void)root;
-	(void)name;
-	(void)value;
-	(void)overwrite;
-
-	return __set_errno(ENOSYS, -1);
-}
-
 ssize_t path_resolution(int dfd, const char *path, char *buf, size_t bufsiz,
 			int atflags)
 {